ADC: just an extranction of slow ADC code to separate file
